Finish frame embedding
