+++ compositor_chromium/src/v8/object/document.rs
@@ -1,5 +1,5 @@
-use log::error;
 use crate::cef::{V8ContextEntered, V8Element, V8Object, V8ObjectError, V8String, V8Value};
+use log::error;
 
 pub struct V8Document(pub V8Object);
